Check for AODH host before adding alarm_connection
Currently the os_ceilometer role will configure ceilometer.conf to
include a line in the [databases] section for alarm_connection no matter
what, and it will default to looking for the aodh user in mongodb
serving at localhost. If the user installs ceilometer but not aodh, this
will break the ceilometer service as the API will not start up while it
times out looking for the unconfigured database. This patch ensures that
there are hosts configured for AODH before adding that line to the conf
file.

Change-Id: If27e19e7c6bb2d20e7b0093a8d4e29c30952fb14
Closes-bug: #1549426