Fail fast when required secrets are not present
This patch aims to establish a pattern for all OSA roles to adopt in which
required variables are validated as present before any part of the role
executes.

There are currently many roles which will "succeed", i.e. their tasks will
complete without raising an error but the resulting configuration created by
the role is either insecure or otherwise fundamentally broken.

Change-Id: Ibc7c770d31bd3fcf340209e75581295dec8fb324