[DOCS] Move ceilometer configs from developer docs
This commit moves the ceilometer configs in the developer docs
over to the role. It also cleans up README.rst and index.rst to
support the new tree structure for role docs.

As a note, some of the role docs(like this one), do a literalinclude
on files in the integrated repo. These references need to be fixed
to reference the link since the included file will no longer exist
in a guaranteed location that can be referenced with relative paths.

Change-Id: Icfaa49c1876b633adf6b61881006c676b1a57cdc
Signed-off-by: alextricity25 <miguel.cantu@rackspace.com>