Disable list_extend in config_template
This allow ceilometer pipelines to be fully controlled with config_template.
Previously a ceilometer pipeline would extend a nested list instead of
redefining it which can have undesirable effects. This change disables the
default list_extend property within the config_template allowing a deployer
to fully redefine a given pipeline.

Change-Id: Ibe64254dcbca99b6b720146d2d4f2a25e22236a3
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>