Reinitialize venv upon install
Reinitializes (copies python, etc binaries) into the venv when
dropping a new venv into place. This is needed because the Python
binary packaged with the venv may not match the Python running on
the host it is being installed to. (ie. in the case of a Xenial
repo container and a Trusty target host.)

Change-Id: Ic49729f7e4f4b94311e5264efd49b2574a71b90b
Partial-Bug: #1637509
(cherry picked from commit 2951230035c61b01e27998918f390c3ecedd5542)