Remove Trusty support from os_ceilometer role
As a part of removing Trusty from OpenStack-Ansible we shall aim to
remove this from all the roles.

Testing has already been removed for Trusty in the integrated build and
all individual repositories on master (Ocata), as such we can now go
ahead and remove the support within the roles.

Change-Id: Ia564599a02eeafecc5614271fe4afbb1eb812aea
Implements: blueprint trusty-removal