Remove unrequired templates in favour of upstream
This patch updates the templated files we have failed to update, and
will pull these from upstream in future. This will ensure the version of
Ceilometer matches the conf files that are used.

pipeline.yaml is still carried because we override some vars in that
file. This is a TO-DO for later, but this patch will reduce the amount
of files we need to carry in templates.

Change-Id: I64fa7af1ca7a58214b6c302131964b819aad4915