Add release note for git_config_lookup_location
Although the feature doesn't impact upgrades or users experiences, we
now expose the option to adjust the git repository that is required to
retrieve the configuration files.

This feature is important for deployers with no access out (except for
from the repo host for example), so we should add a releasenote to bring
attention to this.

Change-Id: I0f0ab474a7ac86fd0273949b54be6c0ad4a89721
(cherry picked from commit 37890ba89a835476bcca3b678133191c365258ed)