Revert "Pin gnocchiclient to comply with optional deps"
This reverts commit f8995cf0843663677d9b876dd06d21956e9271b3.

Adding the constraints to the requirements list prepared by the
py_pkgs lookup is ineffective because it's overidden by the
constraints fed to the wheel build process.

Instead the correct solution is to ensure that these constraints
are provided in the constraints data set. This has been done
in https://review.openstack.org/466660

Change-Id: Ia499fd56440319b3dbd2c3b9ed0ff452943abf8c