Add support for the openSUSE Leap distributions
Add support for the openSUSE Leap distributions. Nothing special is
required for this except for adding the appropriate distro variables
file and also update the zypper cache before package installation.
Moreover, for tests to pass, we need to use the OBS repository to
install the MongoDB packages since they are not provided in the default
openSUSE distribution repositories for Leap < 42.3

Change-Id: Ifa3968f7f578174cf12737c758fa2bca30393fb0