Convert role to use a common systemd service role
This removes the systemd service templates and tasks from this role and
leverages a common systemd service role instead. This change removes a
lot of code duplication across all roles all without sacrificing features
or functionality. The intention of this change is to ensure uniformity and
reduce the maintenance burden on the community when sweeping changes are
needed.

Change-Id: I128bc9736c1a460c161314eeaa6759de5ceba977
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>