Only replace python when re-initializing the venv
We reinitialize the venv to ensure that the right version of
python is in the venv, but we do not want virtualenv to also
replace pip, setuptools and wheel so we tell it not to. If we
don't do this then virtualenv will install the latest available
version, which is not what we want.

Change-Id: I3c589dd4c105bb247f06479d3881ffd68e699224
Partial-Bug: #1764470
(cherry picked from commit 07d9cc32bed99f4f7613d5f7f296f9c4a60313ed)