Execute service setup against a delegated host using Ansible built-in modules
In order to reduce the packages required to pip install on to the hosts,
we allow the service setup to be delegated to a specific host, defaulting
to the deploy host. We also switch as many tasks as possible to using the
built-in Ansible modules which make use of the shade library.

The 'virtualenv' package is now installed appropriately by the openstack_hosts
role, so there's no need to install it any more. The 'httplib2' package is a
legacy Ansible requirement for the get_url/get_uri module which is no longer
needed. As there are no required packages left, the task to install them is
also removed.

The openstack_openrc role is now executed once on the designated host, so
it is no longer necessary as a meta-dependency for the role.

Ceilometer no longer has an API service, so the service setup task for it is
removed as it is unnecessary, along with the related variables which are no
longer used.

Depends-On: https://review.openstack.org/579233
Depends-On: https://review.openstack.org/579959
Change-Id: I4072acc1770432526a8bc26ebb5833b6fdd61f0a