Switch to simple service test instead of tempest
It appears that ceilometer itself has no tests now that there
is no API, so we switch to simply verifying if the services
are up.

Distro installs for xenial are not working, which is not
surprising given that the service is using queens packages.
We therefore disable the test for follow up in another patch.

Change-Id: Ia06438c22430c3432f6dff523c22ec3cd7bdddf8
(cherry picked from commit 854569cdb86b51d07dc8994254aee5dc65c1bd5a)