<service>_ceilometer_enabled defaults based on services existence
Previously common checks were used in group_vars, and they were checking
if ceilometer exists. But ceilometer doesn't have such group_var.
As a result service thought, that ceilometer is enabled, and was sending
notifications to message queue.
These messages were accumulating, as the consumer was absent and
knew nothing about this service and it's queue.
Now this situation should be fixed.

Change-Id: I29b7f679eed1249f5367c42a2898804f1a436cbc