Enable overriding the service setup host python interpreter
In order to enable the service setup host python interpreter to
be changed easily, we make it a variable. This will be useful
when someone sets the service setup host to be the utility
container, because we'll be able to set this var by default.

Change-Id: I17b580fa357154214da93a473213e7f37980a45d
(cherry picked from commit 40baff22aa65beacc46529067be707eedc5af8d5)