Use the distro package for libvirt-python
The libvirt-python library has C bindings which are very particular
about the version of its companion package (libvirt). To ensure
things run smoothly for stable releases, we opt to use the distro
packages for these and symlink the appropriate library files and
binding into the venv.

This approach has been used successfully for the ceph and libvirt
python bindings in other roles.

Note:
This is not a clean cherry-pick due to the change in the way the
python packages are installed in Stein.

Change-Id: I23bf430087af8b16a3d77d96f6ceb3d7f6d8a9d5
(cherry picked from commit 6d2ad6a0042a07ec9e7eb330b86534539cc37980)